requirements.txt
requests
pandas
numpy
heapq
flask
pickle
scikit-surprise
google-cloud-storage
gunicorn